KAN-103 rename: planner types move to app/types
